update variant getting function
